Добавил вывод полей; разделил классы тестов по разным файлам
