+++ ObjectPrinting/ObjectPrinting.csproj
@@ -90,7 +90,8 @@
     <Compile Include="PropertySerializingConfig.cs" />
     <Compile Include="PropertySerializingConfigExtensions.cs" />
     <Compile Include="Tests\ClassWithCircularReference.cs" />
-    <Compile Include="Tests\ObjectPrinter_Tests.cs" />
+    <Compile Include="Tests\ObjectPrinterAcceptanceTests.cs" />
+    <Compile Include="Tests\ObjectPrinter_Should.cs" />
     <Compile Include="Tests\Person.cs" />
     <Compile Include="Tests\ClassWithStrings.cs" />
   </ItemGroup>
