+++ ObjectPrinting/Tests/ClassWithStrings.cs
@@ -2,7 +2,7 @@
 {
     public class ClassWithStrings
     {
-        public string s1 { get; set; }
-        public string s2 { get; set; }
+        public string s1;
+        public string s2;
     }
 }
